Base architecture
The Gradle modules created are:
:feature-market - represents the UI.

:core - will be used for separation of concerns between the UI and the
data modules.

:data-repository - will receive the data from the :data-local Gradle
module. With this data, flows are returned to stream the data to the UI.

:data-remote - will retrieve the data from the backend.

:data-local - will store and retrieve data from the local database.

The main reason to create all the different Gradle modules is for
scalability, separation of concerns, better testing among other benefits.

Relation between the modules:
:app the entry point of the app depends on :feature-users
:feature-marketplace the UI depends on :core
:data-repository depends on :data-remote, data-local and :core
:core module acts as the abstraction module. That is, the UI doesn't
know anything about how the data is retrieved.

The dependency inversion principle is follow here by depending on
interfaces instead of concrete functions.